refactor: 트렌딩 set 자료구조로 저장하게 변경, 키 삭제 기능 추가
