+++ src/main/java/net/pointofviews/movie/batch/trending/TMDbTrendingMovieStep.java
@@ -3,10 +3,7 @@ package net.pointofviews.movie.batch.trending;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.pointofviews.movie.domain.Movie;
-import org.springframework.batch.core.ExitStatus;
 import org.springframework.batch.core.Step;
-import org.springframework.batch.core.StepExecution;
-import org.springframework.batch.core.StepExecutionListener;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.step.builder.StepBuilder;
 import org.springframework.context.annotation.Bean;
@@ -33,21 +30,7 @@ public class TMDbTrendingMovieStep {
                 .reader(trendingMovieReader)
                 .processor(trendingMovieProcessor)
                 .writer(trendingMovieWriter)
-                .listener(new StepExecutionListener() {
-                    @Override
-                    public ExitStatus afterStep(StepExecution stepExecution) {
-                        long savedMoviesCount = trendingMovieWriter.countTrendingMovies();
-
-                        if (savedMoviesCount >= 20) {
-                            log.info("트렌딩 영화 배치 완료: {} 개 저장", savedMoviesCount);
-                            stepExecution.setExitStatus(ExitStatus.COMPLETED);
-                            return ExitStatus.COMPLETED;
-                        }
-
-                        log.info("현재 저장된 트렌딩 영화 수: {}", savedMoviesCount);
-                        return stepExecution.getExitStatus();
-                    }
-                })
+                .listener(trendingMovieWriter)
                 .build();
     }
 }
